# apps/backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ErrorResponse:
    properties:
      detail:
        description: optional internal error
        example: validation failed on field 'email'
        type: string
      error:
        description: user-friendly message
        example: Invalid request
        type: string
    type: object
...
  notification.CreateNotificationRequest:
    properties:
      message:
        type: string
      type:
        $ref: '#/definitions/notification.NotificationType'
      user_id:
        type: string
    type: object
  notification.Notification:
    properties:
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
      sent_at:
        type: string
      status:
        $ref: '#/definitions/notification.NotificationStatus'
      type:
        $ref: '#/definitions/notification.NotificationType'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  notification.NotificationStatus:
    enum:
    - pending
    - sent
    - failed
    - read
    type: string
    x-enum-varnames:
    - Pending
    - Sent
    - Failed
    - Read
  notification.NotificationType:
    enum:
    - email
    - sms
    - push
    - system
    type: string
    x-enum-comments:
      System: for in-app or placeholder
    x-enum-descriptions:
    - ""
    - ""
    - ""
    - for in-app or placeholder
    x-enum-varnames:
    - Email
    - SMS
    - Push
    - System
  notification.UpdateNotificationStatusRequest:
    properties:
      status:
        $ref: '#/definitions/notification.NotificationStatus'
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: This is the API for fastabiz operations.
  title: FastaBiz API
  version: "1.0"
paths:
...
  /notifications/{id}/read:
    patch:
      consumes:
      - application/json
      description: Update notification status to "read" by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to update notification
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark a single notification as read
      tags:
      - notifications
  /notifications/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update notification status by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update request
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/notification.UpdateNotificationStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Notification not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to update notification
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a notification's status (e.g. mark as sent or read)
      tags:
      - notifications
  /notifications/all_my_notifications/{id}:
    get:
      description: Get all notifications belonging to a user (optionally filtered
        by status)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Notification status (pending, sent, failed, read)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/notification.Notification'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List notifications by user
      tags:
      - notifications
  /notifications/all_pending_notifications:
    get:
      description: Get a list of notifications not yet sent
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/notification.Notification'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all pending notifications
      tags:
      - notifications
  /notifications/create:
    post:
      consumes:
      - application/json
      description: Create a new notification with user_id, message, etc.
      parameters:
      - description: User Input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/notification.CreateNotificationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/notification.Notification'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Failed to create notification
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new notification
      tags:
      - notifications
  /notifications/mark_all_as_read/{id}:
    patch:
      description: Mark all unread notifications for a given user ID as "read"
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to mark notifications as read
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark all user notifications as read
      tags:
      - notifications
...
schemes:
- http
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"